clean up docstrings and file namings
